+++ R/utils_model_family.R
@@ -4,6 +4,11 @@ get_glm_family <- function(fit) {
   # get model class
   mc <- get_predict_function(fit)
 
+  # for gam-components from gamm4, add class attributes, so family
+  # function works correctly
+  if (any(mc == "gam") && !any(c("glm", "lm") %in% mc))
+    class(fit) <- c(class(fit), "glm", "lm")
+
   # do we have glm? if so, get link family. make exceptions
   # for specific models that don't have family function
   if (any(mc %in% c("lme", "plm", "gls", "truncreg"))) {
@@ -30,7 +35,7 @@ get_glm_family <- function(fit) {
   } else {
     # "lrm"-object from pkg "rms" have no family method
     # so we construct a logistic-regression-family-object
-    if (mc %in% c("lrm", "polr", "logistf", "clm"))
+    if (mc %in% c("lrm", "polr", "logistf", "clm", "multinom"))
       faminfo <- stats::binomial(link = "logit")
     else
       # get family info
@@ -42,18 +47,21 @@ get_glm_family <- function(fit) {
   }
 
   # create logical for family
-  binom_fam <-
+  binom_fam <- any(
     fitfam %in% c("binomial", "quasibinomial", "binomialff") |
     sjmisc::str_contains(fitfam, "binomial", ignore.case = TRUE)
+  )
 
-  poisson_fam <-
+  poisson_fam <- any(
     fitfam %in% c("poisson", "quasipoisson") |
     sjmisc::str_contains(fitfam, "poisson", ignore.case = TRUE)
+  )
 
-  neg_bin_fam <-
+  neg_bin_fam <- any(
     sjmisc::str_contains(fitfam, "negative binomial", ignore.case = T) |
     sjmisc::str_contains(fitfam, "nbinom", ignore.case = TRUE) |
     sjmisc::str_contains(fitfam, "neg_binomial", ignore.case = TRUE)
+  )
 
 
   list(